Avoid duplicates in default_include_directories.
When the user does not specify --prefix, the default
value often matches one of the hardcoded defaults
in default_include_directories.
This results in a duplicate entry in default_include_directories.
This in turn results in make doing extra work looking for
included file in the same directory twice.
sv 60378.