Enhance implicit search logging.
Consider this rule
%.o: %.c
    $(CC) -o $@ -c $<

When the user wants to make hello.o and hello.c is missing, make gives the
following diagnostics.

$ ls
makefile
$ cat makefile
%.o: %.c; $(CC) -o $@ -c $<
$ ~/src/gmake/make/m64/make -R --debug=implicit hello.o |tail +9
Updating goal targets....
 File 'hello.o' does not exist.
 Looking for an implicit rule for 'hello.o'.
 Trying pattern rule with stem 'hello'.
 Trying implicit prerequisite 'hello.c'.
 Trying pattern rule with stem 'hello'.
 Trying implicit prerequisite 'hello.c'.
 Looking for a rule with intermediate file 'hello.c'.
  Avoiding implicit rule recursion.
 No implicit rule found for 'hello.o'.
Must remake target 'hello.o'.
make: *** No rule to make target 'hello.o'.  Stop.
$

The bottom line in this diagnostics is the "No rule found" message.
While the issue is that hello.c is missing.
And the user stares at the makefile and says
"How cannot make find the rule? Here it is."

Preferably make should tell which rule it is trying and why it is rejecting it
and which prerequisite make is trying and what the outcome is.

Changes in the implicit rule search logging:
1. The messages which refer to a rule print the first line
    (the targets and prerequisites) of the rule.
2. A new message tells when a rule is rejected, along with the reason.
3. The 'Looking for an implicit rule...' message is printed for every
    prerequisite, rather than the top level target only.
4. "Trying harder" message is printed, when intermediate prerequisites are
    allowed.
5. The 'No rule found...' and 'Found implicit rule...' messages are printed for
    every prerequisite, rather than the top level target only.
6. "Ought to exist...", "Found..." or "Not found..." message is printed for
    each prerequisite.

This is the diagnostics of the same example above with enhanced logging.

$ ~/src/gmake/make/l64/make -R --debug=implicit hello.o |tail +9
Updating goal targets....
 File 'hello.o' does not exist.
 Looking for an implicit rule for 'hello.o'.
 Trying pattern rule '%.o: %.c' with stem 'hello'.
 Trying implicit prerequisite 'hello.c'.
 Not found 'hello.c'.
 Trying harder.
 Trying pattern rule '%.o: %.c' with stem 'hello'.
 Trying implicit prerequisite 'hello.c'.
 Looking for a rule with intermediate file 'hello.c'.
  Avoiding implicit rule recursion for rule '%.o: %.c'.
  No implicit rule found for 'hello.c'.
 Rejecting rule '%.o: %.c' due to impossible prerequisite 'hello.c'.
 No implicit rule found for 'hello.o'.
Must remake target 'hello.o'.
make: *** No rule to make target 'hello.o'.  Stop.