Remake if the mtimes of a target and its prerequisite match.
It is possible for a file to be updated, but mtime be equal to
this_mtime. This could happen when a program creates a file and then
immediately proceeds and runs make and make updates a prerequisite of
the file just created. If this happens fast enough, then both the file
and its prerequisite have the same mtime.

* remake.c (check_dep): Remake a target if its prerequisite was created even if
the mtimes of the target and the prerequisite match.