[SV 62206] Fix % substitution in pattern rules.
Support multiple patterns in each prerequisite, during second expansion of
pattern rule prerequisites.
E.g. given makefile

.SECONDEXPANSION:
all: hello.x
%.x: $$(wordlist 1, 99, %.1 %.2) $$(wordlist 1, 99, %.3 %.4); $(info $@ from $^)

this commit implements substitution of '%' with 'hello' in '%.2' and '%.4'.

* src/rule.c (snap_implicit_rules): During max_pattern_dep_length calculation
count each % as 5 bytes when second expansion is enabled to preallocate
enough memory for $(*F).
* src/implicit.c (pattern_search): When second expansion is enabled,
substitute the first % of each prerequisite word with *$ or $(*F).
* tests/scripts/features/patternrules_stem: Test that make correctly computes
stem and substitutes % for computed stem in pattern rules with and without
second expansion and with and without last slash.