[SV xxx] Fix % substitution in static pattern rules.
Keep the second and following % intact in a prerequisite of a
static pattern rule second expansion is enabled.
E.g. given makefile

.SECONDEXPANSION:
all: hello.x
hello.x: %.x: $$(wordlist 1, 99, %_%.1 %_%.2); $(info $@ from $^)

this commit keep the second '%' in each prerequisite intact.

* src/file.c (expand_deps): When second expansion is enabled,
keep the second and following % intact in a prerequisite of a static pattern
rule.
* tests/scripts/features/statipattrules_stem: Test that make correctly computes
stem and substitutes % for computed stem in static pattern rules with and
without second expansion and with and without last slash.