[SV 62278] Fix % substitution in static pattern rules.
Keep the second and following % intact in a prerequisite of a
static pattern rule when second expansion is enabled.
E.g. given makefile

.SECONDEXPANSION:
all: hello.x
hello.x: %.x: $$(wordlist 1, 99, %_%.1 %_%.2); $(info $@ from $^)

this commit keep the second and following '%' in each prerequisite intact.

* src/file.c (expand_deps): When second expansion is enabled,
keep the second and following % intact in a prerequisite of a static
pattern rule.
(expand_deps): Fix a leak of of struct dep.
* tests/scripts/features/statipattrules_stem: Test that make correctly
computes stem and substitutes % for the computed stem in static pattern
rules with and without second expansion and with and without last slash.