[SV 62734] Have ASAN ignore leaks.
There are known memory leaks.
When make is build with ASAN checks enabled, those leaks fail the tests.
This patch sets ASAN env to ignore leaks. This simply makes it easier to
run the test suite with ASAN enabled builds of make.
Once (if ever) those known leaks are removed this commit can be reverted.

* tests/run_make_tests.pl: Have the test suite set ASAN env to ignore leaks.