[SV 63417] Fix global .NOTINTERMEDIATE.
Fix .NOTINTERMEDIATE without prerequisites to really cause all targets to be
treated as not intermediate.

* filedef.h: Declare extern no_intermediates.
* file.c: Remove static from no_intermediates.
(remove_intermediates): Check no_intermediates.
* implicit.c (pattern_search): For a file found by implicit search set
file->notintermediate if the pattern has notintermediate set or
no_intermediates is set.
* remake.c (update_file_1): Avoid setting file->secondary for a pre-existing
file if file->notintermediate is set or no_intermediates is set. The check for
no_intermediates here is redundant, but won't hurt. Let it be in case make code
changes such that its presence matters.
* ../tests/scripts/targets/NOTINTERMEDIATE: Fix a test.