[SV 63821] Move setup of default suffix rules after parsing makefiles.
This ensures that when built-in rules are disabled in the makefile, but
suffixes are added to .SUFFIXES, make exits with the 'No rule...' error
message.
This also relieves make from creating and then destroying file objects for
default suffix rules and keeps hash table "files" free of the related entries
when built-in rules are disabled in the makefile.

* src/main.c (main): Install default suffix rules after parsing makefiles.
* src/default.c (install_default_suffix_rules): Install the default suffix
rule only if there is no user defined rule.
* tests/scripts/features/suffixrules: Add tests.