[SV 63840] Prevent match-anything rules from matching default suffixes.
Match-anything rules are not supposed to apply to files with known suffixes.
This commit ensures that this restriction holds even when built-in rules are
disabled.

* src/makeint.h: Declare match_default_suffixes.
* src/default.c (match_default_suffixes): Define match_default_suffixes.
* src/implicit.c (pattern_search): Call match_default_suffixes to check if a
file has a known suffix.
* src/rule.c (convert_to_pattern): Do not convert the default suffixes to dummy
pattern rules. With introduction of match_default_suffixes, there is no need
for these pattern rules. Keeps pattern_rules smaller and reduces the amont of
rules pattern_search has to go through.
* doc/make.texi: Update the manual.
* tests/scripts/features/matchanyrules: Add new tests.
* tests/scripts/features/suffixrules: Ditto.