[SV 63840] Prevent match-anything rules from matching default suffixes.
Match-anything rules are not supposed to apply to files with the default
or user defined suffixes. Ensure that this restriction holds even when
built-in rules are disabled.

* src/default.c (set_default_suffixes): Define default_suffixes with extern
linkage.
(set_default_suffixes): Enter the default suffixes whether the built-in rules
are enabled or not.  When the built-in rules are disabled, the default
suffixes are still needed to disqualify match-anything rules.
* src/read.c (record_files): Reset file->builtin for each user defined suffix.
This makes a difference when the user defines suffixes that match the default
suffixes.
* src/rule.c (convert_default_suffixes): Declare extern default_suffixes.
(convert_default_suffixes): Convert all default suffixes from array
default_suffixes to dummy pattern rules.
(convert_to_pattern): Call convert_default_suffixes.
(convert_to_pattern): When built-in rules are disabled, convert suffix rules
only for the suffixes defined in the makefile.
* tests/scripts/features/matchanyrules: Add new tests.
* tests/scripts/features/suffixrules: Ditto.