[SV 63856] Fix pruning of double colon rules.
* src/remake.c (update_file): Fix premature pruning of double colon rules.

Given this makefile

A::; @echo A-1 && sleep 1 && echo A-1 done
A::; @echo A-2 && sleep 1 && echo A-2 done
A::; @echo A-3 && sleep 1 && echo A-3 done
B::; @echo B-1 && sleep 1 && echo B-1 done
B::; @echo B-2 && sleep 1 && echo B-2 done
B::; @echo B-3 && sleep 1 && echo B-3 done

And running
$ make -j8 A .WAIT B
all 3 recipes for A should get started and complete sequentially before the 1st
recipe for B gets started and then all 3 recipes for B should get started and
complete sequentially.
The change in update_file is needed, because update_file prunes A even
when the current A recipe is already done and the recipe of the next double
colon rule of the same target should be started. This incorrect pruning
causes update_goal_chain to run a recipe of the next goal, even though there
are still not started recipes of this goal.
This patch modifies the pruning criteria such that for double colon rules
pruning only takes place if the current recipe is still in progress.

* tests/scripts/targets/WAIT: Test .WAIT with double colon rules.