[SV 64822, SV 36486] Fix appending to a pattern specific variable.
Appending to a pattern specific variable produces an incorrect value in the
presence of a command line definition or an env override of the variable.
Also, fix pattern/target-specific appending to a variable with origin override.

The following takes place.
1. At parse time record_target_var sets the value of a pattern specific
variable to the value defined on command line or to the value of the env
override.
2. Later, at build time, recursively_expand_for_file appends this value of the
variable (set in record_target_var) to the command line value again, regardless
of the origin of the variable.

This patch modifies recursively_expand_for_file to avoid appending, unless the
origin of the variable beats or equals the origin of one of the parent
definitions of this variable.

* src/variable.c (do_variable_definition): Avoid merging a
pattern-specific variable with the parent definition when a command
line or env override is present.
* src/expand.c (recursively_expand_for_file): Avoid appending to a
pattern-specific variable, unless the origin of this pattern-specific variable
beats or equals the origin of one of the parent definitions of this variable.
* tests/scripts/variables/append: Add tests.

Reported by Rob <robw9739@gmail.com>,
Brian Vandenberg <phantall@gmail.com>,
Markus Oberhumer <markus@oberhumer.com>.