[SV xxx] Sort print-targets output.
The hashing routine hashes differently on big and little endian machines.
This causes targets to be printed in some (pseudo-random) order on little
endian and some other (also pseudo-random) order on big endian.
Sort targets in the order as they are defined in the makefile.

* src/file.c (compare_files_by_loc): Add targets sorting routine.
(print_targets): Sort the targets.
* tests/scripts/options/print-targets: Update the test to expect sorting.