[sv 65588] Handle an unterminated quote in shellflags.
In one-shell mode with shellflags that contains an unterminated quote, pass the
recipe along with shellflags to default shell to have the shell deal with the
quote. With this change make's handling of an unterminated .SHELLFLAGS in
one-shell mode is the same as when .ONESHELL is not specified.

* src/job.c (construct_command_argv_internal): When shellflags contains
an unterminated shellflags, run default shell in oneshell mode.
* tests/scripts/variables/SHELLFLAGS: Add tests.